Aura Points/ml/requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
fastapi==0.104.1
uvicorn[standard]==0.24.0

